removed VMBridge
